updates OS X project; fixes various OS X compiler/run-time fixes (still doesn't run); switches to 64 bit
Former-commit-id: fce9e735ef46a5ea810ecb192e21ed99f8dde7f2

+++ xcode4/crag/SDLmain/SDLMain.m
@@ -315,7 +315,7 @@ static void CustomApplicationMain (int argc, char **argv)
     NSRange localRange;
     NSString *result;
 
-    bufferSize = selfLen + aStringLen - aRange.length;
+    bufferSize = (unsigned int)(selfLen + aStringLen - aRange.length);
     buffer = (unichar *)NSAllocateMemoryPages(bufferSize*sizeof(unichar));
     
     /* Get first part into buffer */
